Move validation/parsing exception classes to dbt_semantic_interfaces
