+++ dbt_semantic_interfaces/dbt_semantic_interfaces/objects/base.py
@@ -7,7 +7,7 @@ from typing import Any, Callable, ClassVar, Generic, Iterator, TypeVar
 
 from pydantic import BaseModel, root_validator
 
-from metricflow.errors.errors import ParsingException
+from dbt_semantic_interfaces.errors import ParsingException
 from dbt_semantic_interfaces.parsing.yaml_loader import ParsingContext, PARSING_CONTEXT_KEY
 
 # Type alias for the implicit "Any" type used as input and output for Pydantic's parsing API
