+++ dbt_semantic_interfaces/dbt_semantic_interfaces/objects/common.py
@@ -4,7 +4,7 @@ import re
 
 from typing import Optional
 
-from metricflow.errors.errors import ParsingException
+from dbt_semantic_interfaces.errors import ParsingException
 from dbt_semantic_interfaces.objects.base import HashableBaseModel
