+++ dbt_semantic_interfaces/dbt_semantic_interfaces/objects/constraints/where.py
@@ -5,7 +5,7 @@ import re
 from typing import List, Optional, Dict, Any
 
 from mo_sql_parsing import parse as mo_parse
-from metricflow.errors.errors import ConstraintParseException
+from dbt_semantic_interfaces.errors import ConstraintParseException
 from dbt_semantic_interfaces.objects.base import (
     HashableBaseModel,
     PydanticCustomInputParser,
