+++ dbt_semantic_interfaces/dbt_semantic_interfaces/transformations/add_input_metric_measures.py
@@ -1,6 +1,6 @@
 from typing import Set
 
-from metricflow.errors.errors import ModelTransformError
+from dbt_semantic_interfaces.errors import ModelTransformError
 from dbt_semantic_interfaces.objects.metric import MetricType, MetricInputMeasure
 from dbt_semantic_interfaces.objects.user_configured_model import UserConfiguredModel
 from dbt_semantic_interfaces.transformations.transform_rule import ModelTransformRule
