+++ dbt_semantic_interfaces/dbt_semantic_interfaces/transformations/convert_count.py
@@ -1,5 +1,5 @@
 from dbt_semantic_interfaces.objects.aggregation_type import AggregationType
-from metricflow.errors.errors import ModelTransformError
+from dbt_semantic_interfaces.errors import ModelTransformError
 from dbt_semantic_interfaces.objects.user_configured_model import UserConfiguredModel
 from dbt_semantic_interfaces.transformations.transform_rule import ModelTransformRule
