+++ dbt_semantic_interfaces/dbt_semantic_interfaces/transformations/proxy_measure.py
@@ -1,6 +1,6 @@
 import logging
 
-from metricflow.errors.errors import ModelTransformError
+from dbt_semantic_interfaces.errors import ModelTransformError
 from dbt_semantic_interfaces.objects.metric import Metric, MetricInputMeasure, MetricType, MetricTypeParams
 from dbt_semantic_interfaces.objects.user_configured_model import UserConfiguredModel
 from dbt_semantic_interfaces.transformations.transform_rule import ModelTransformRule
